Reworking crate structure
