--- dtor-proc-macro/Cargo.toml
@@ -1,15 +1,0 @@
-[package]
-name = "dtor-proc-macro"
-version = "0.0.13"
-authors.workspace = true
-edition = "2021"
-description = "proc-macro support for the dtor crate"
-license.workspace = true
-repository.workspace = true
-readme = "README.md"
-
-[features]
-default = []
-
-[lib]
-proc-macro = true
