--- link-section-proc-macro/README.md
@@ -1,3 +1,0 @@
-# link-section-proc-macro
-
-A proc macro for the `link-section` crate.
