+++ linktime-proc-macro/src/xx3.rs
